+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Added lossless container-level metadata sanitization for JPEG and WebP images.
+
 ## [2.0.0] - 2026-08-18
 
 ### Changed